Moving business logic to service layer
Renamed method name to make it more intuitive
