+++ src/main/java/com/birichani/code/restapi/api/VideoApiController.java
@@ -1,7 +1,5 @@
 package com.birichani.code.restapi.api;
 
-import com.birichani.code.restapi.constant.ErrorMessage;
-import com.birichani.code.restapi.constant.InfoMessage;
 import com.birichani.code.restapi.model.Video;
 import com.birichani.code.restapi.repository.VideoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,6 +7,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import com.birichani.code.restapi.service.VideoService;
 
 import java.util.List;
 
@@ -16,31 +15,20 @@ import java.util.List;
 @RestController
 public class VideoApiController {
 	private final VideoRepository videoRepository;
+	private final VideoService videoService;
 	
 	@Autowired
-	 public VideoApiController(VideoRepository videoRepository) {
+	 public VideoApiController(VideoRepository videoRepository, VideoService videoService) {
 		this.videoRepository = videoRepository;
+		this.videoService = videoService;
 	}
 	
 	@GetMapping("/{videoTopic}")
-	public String video(@PathVariable String videoTopic) {
-		String message;
-		
-		switch (videoTopic.toLowerCase()) {
-			case "python":
-				message = InfoMessage.PYTHON_RESPONSE_MESSAGE;
-				break;
-			case "java":
-				message = InfoMessage.JAVA_RESPONSE_MESSAGE;
-				break;
-			default:
-				message = ErrorMessage.VIDEO_TOPIC_SELECTION_MESSAGE;
-		}
-		
-		return message;
-}
+	public String getVideosByTopic(@PathVariable String videoTopic) {
+		return videoService.getVideosByTopic(videoTopic);
+	}
 	@GetMapping()
-	public List<Video> getAllvideos(){
+	public List<Video> getAllVideos(){
 		return videoRepository.findAll();
 	}
 }
